requirements.txt
pandas
openpyxl
regex
streamlit
python-docx
PyPDF2
textract
pytesseract
pdf2image
